refactor: useAuthProvider를 useAuth 에서 분리
